requirements.txt
pandas==2.2.1
numpy==1.26.4
streamlit==1.32.2
google-generativeai==0.4.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
